+++ golang/lib/fldsort/fldsort.go
@@ -11,7 +11,6 @@ import (
 
 const (
 	toolname = "xsvutils_sort"
-	source   = "/Users/yusuke/repo/xsvutils/golang/lib/fldsort/in.txt"
 )
 
 func writeRow(r string, w *bufio.Writer) error {
@@ -280,6 +279,7 @@ func FieldSort(
 	}
 	sc := bufio.NewScanner(f)
 	defer file.Close()
+	defer os.Remove(file.Name())
 
 	i := 0
 	for sc.Scan() {